refactor(sim-plugin): type control-plane numbers by their role
Ports, sessions, multicast groups, replication ids, node handles and
extern array sizes and indices are all `usize`, converted from P4 values
at the use site; a value past the machine word surfaces as
`ExternError::MachineWord`, and an index at or past the array length
stays a no-op as the P4 extern documentation specifies. STF integers
parse through one generic `parse_int` over `strtoint`, so ports parse
as `usize` and values as `i128` instead of everything as `i64`.

+++ p4spec-rust/src/sim-plugin/psa/mirror.rs
@@ -1,4 +1,4 @@
 use std::collections::BTreeMap;
 
 /// Mirror table: clone session id to multicast group id
-pub type Table = BTreeMap<i64, i64>;
+pub type Table = BTreeMap<usize, usize>;
